test(message-buffer): refactor integration test structure
